[code] argparse: boost::po -> taywee/args
Modification of the arguments parsing library.

The CLI would not have changed at all (unless non-standard arguments
passing has been used before, such as using long options with a single
dash).

This commit is quite raw and does not show the benefits from using
taywee/args. The API is cleaner and easier to use, and may allow many
improvements in the long run:
- BASH COMPLETION!
- defining ad hoc types for parsing complex parameters
- defining a hierarchy of exceptions to improve error logging
- have a generated clean help format, with groups of parameters

This library may also be used for Batsim later on.